add v3 and v4 models
v3 model uses the full dataset, in order to see how it performs in real
world situations compared to the other models, the thinking is that
since this is a time series, we want to train on the most recent data.
the v0-v2 models all train on 2020-2023 data, and test on 2023-2025
data. Which means for predicting 2025 data, the training data is out of
date.

v4 is the same as v3, but it will train longer based on patience
level=3, rather than a fixed interval